# app/static/swagger.yaml
openapi: 3.0.0
info:
  version: "1.0.0"
  title: "Lunch API"
  description: "This is a Lunch API"
servers:
  - url: http://0.0.0.0:5000
...
paths:
...
  /api/auth/registration:
    post:
      tags:
        - "Authentication"
      summary: "Registration of a new employee"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserIn'
      responses:
        '201':
          description: "Successful Operation"
          content:
            application/json:
              example:
                id: "5"
                access_token: "{access_token}"
                refresh_token: "{refresh_token}"
        '400':
          description: "Wrong input data"
          content:
            application/json:
              example:
                message: "Please, specify 'firstname', 'lastname', 'email', 'password' and 'is_admin'"
        '404':
          description: "Wrong input email"
          content:
            application/json:
              example:
                message: "Email {email} already used"
        '500':
          description: "Error occurred"
          content:
            application/json:
              example:
                message: "Something went wrong while creating"
                error: "{error}"
  /api/auth/login:
    post:
      tags:
        - "Authentication"
      summary: "Login as an employee"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LogIn'
      responses:
        '201':
          description: "Successful Operation"
          content:
            application/json:
              example:
                message: "Logged in as {user.firstname, user.lastname, user.email}"
                access_token: "{access_token}"
                refresh_token: "{refresh_token}"
        '400':
          description: "Wrong input data"
          content:
            application/json:
              example:
                message: "Please, provide 'email' and 'password' in body"
        '404':
          description: "Wrong input email or password"
          content:
            application/json:
              example:
                message: "User with email {email} doesn't exist"
  /api/auth/refresh:
    post:
      security:
        - bearerAuth: []
      tags:
        - "Authentication"
      summary: "Refresh access token"
      description: "This can only be done by the logged in user with refresh token"
      responses:
        '201':
          description: "Successful Operation"
          content:
            application/json:
              example:
                access_token: "{access_token}"
        '401':
          description: "Require authorized user"
          content:
            application/json:
              example:
                msg: "Missing Authorization Header"
        '404':
          description: "Wrong input email"
          content:
            application/json:
              example:
                message: "User with email {email} doesn't exist"
        '422':
          description: "Refresh token required"
          content:
            application/json:
              example:
                msg: "Only refresh tokens are allowed"
  /api/auth/logout-access:
    post:
      security:
        - bearerAuth: []
      tags:
        - "Authentication"
      summary: "Revoke access token"
      description: "This can only be done by the logged in user"
      responses:
        '200':
          description: "Successful Operation"
          content:
            application/json:
              example:
                message: "Access token has been revoked"
        '401':
          description: "Require authorized user"
          content:
            application/json:
              example:
                msg: "Missing Authorization Header"
        '500':
          description: "Error occurred"
          content:
            application/json:
              example:
                message: "Something went wrong while creating"
                error: "{error}"
  /api/auth/logout-refresh:
    post:
      security:
        - bearerAuth: []
      tags:
        - "Authentication"
      summary: "Revoke refresh token"
      description: "This can only be done by the logged in user with refresh token"
      responses:
        '200':
          description: "Successful Operation"
          content:
            application/json:
              example:
                message: "Refresh token has been revoked"
        '401':
          description: "Require authorized user"
          content:
            application/json:
              example:
                msg: "Missing Authorization Header"
        '500':
          description: "Error occurred"
          content:
            application/json:
              example:
                message: "Something went wrong while creating"
                error: "{error}"
...
  /api/choices/:
    post:
      security:
        - bearerAuth: [ ]
      tags:
        - "Choices"
      summary: "Create new choice"
      description: "This can only be done by the logged in employee"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChoiceIn'
      responses:
        '201':
          description: "Successful Operation"
          content:
            application/json:
              example:
                id: "5"
        '400':
          description: "Wrong input data"
          content:
            application/json:
              example:
                message: "Please, specify 'menu_id'"
        '401':
          description: "Require authorized user"
          content:
            application/json:
              example:
                msg: "Missing Authorization Header"
  /api/choices/{id}:
    get:
      tags:
        - "Choices"
      parameters:
        - name: "id"
          in: "path"
          description: "ID of choice"
          required: true
          schema:
            type: "integer"
      summary: "Get choice by id"
      responses:
        '200':
          description: "Successful Operation"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChoiceOut'
        '404':
          description: "Choice not found"
          content:
            application/json:
              example:
                message: "Choice not found"
    patch:
      security:
        - bearerAuth: [ ]
      tags:
        - "Choices"
      summary: "Update choice's information"
      description: "This can only be done by the logged in employee"
      parameters:
        - name: "id"
          in: "path"
          description: "ID of choice"
          required: true
          schema:
            type: "integer"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChoiceIn'
      responses:
        '200':
          description: "Successful Operation"
          content:
            application/json:
              example:
                message: "Updated"
        '401':
          description: "Require authorized user"
          content:
            application/json:
              example:
                msg: "Missing Authorization Header"
        '404':
          description: "Wrong choice's id"
          content:
            application/json:
              example:
                message: "Choice not found"
    delete:
      security:
        - bearerAuth: [ ]
      tags:
        - "Choices"
      summary: "Delete choice"
      description: "This can only be done by the logged in employee"
      parameters:
        - name: "id"
          in: "path"
          description: "Id of the choice"
          required: true
          schema:
            type: "integer"
      responses:
        '200':
          description: "Successful Operation"
          content:
            application/json:
              example:
                message: "Deleted"
        '401':
          description: "Require authorized user"
          content:
            application/json:
              example:
                msg: "Missing Authorization Header"
        '404':
          description: "Wrong choice's id"
          content:
            application/json:
              example:
                message: "Choice not found"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    UserIn:
      type: "object"
      properties:
        firstname:
          type: "string"
          example: "John"
        lastname:
          type: "string"
          example: "Doe"
        email:
          type: "string"
          example: "test@gmail.com"
        password:
          type: "string"
          example: "password"
        is_admin:
          type: "boolean"
          example: true
    LogIn:
      type: "object"
      properties:
        email:
          type: "string"
          example: "test@gmail.com"
        password:
          type: "string"
          example: "password"
...
    ChoiceOut:
      type: "object"
      properties:
        id:
          type: "integer"
          example: 1
        current_day:
          type: "string"
          example: "2022-10-19"
        employee:
          type: "array"
          example: [ ]
        restaurant:
          type: "string"
          example: "McDonald's"
...
    ChoiceIn:
      type: "object"
      properties:
        menu_id:
          type: "string"
          example: "2"